made ApiCaller class for networking calls
